fix: move faqs to component
